+++ tests/test_config_generation.py
@@ -18,7 +18,7 @@ def _assert_yaml_objects(objects, yaml_filename):
     assert objects == desired_objects
 
 
-def _test_configs(folder_name, **kwargs):
+def _test_configs(folder_name, default_registry=None, **kwargs):
     app_dir = path.join('tests', 'configs', folder_name)
 
     kubetools_config = load_kubetools_config(app_dir, **kwargs)
@@ -26,6 +26,7 @@ def _test_configs(folder_name, **kwargs):
     with mock.patch('kubetools.kubernetes.config.job.uuid4', lambda: 'UUID'):
         services, deployments, jobs, cronjobs = generate_kubernetes_configs_for_project(
             kubetools_config,
+            default_registry=default_registry,
         )
 
     k8s_files = listdir(app_dir)
@@ -60,3 +61,6 @@ class TestKubernetesConfigGeneration(TestCase):
 
     def test_multiple_deployments_configs(self):
         _test_configs('multiple_deployments')
+
+    def test_docker_registry_configs(self):
+        _test_configs('docker_registry', default_registry='default-registry')